Move eye_movement workers into separate files
