+++ pupil_src/shared_modules/eye_movement/worker/real_time_buffered_detector.py
@@ -1,0 +1,109 @@
+"""
+(*)~---------------------------------------------------------------------------
+Pupil - eye tracking platform
+Copyright (C) 2012-2019 Pupil Labs
+
+Distributed under the terms of the GNU
+Lesser General Public License (LGPL v3.0).
+See COPYING and COPYING.LESSER for license details.
+---------------------------------------------------------------------------~(*)
+"""
+import typing as t
+import bisect
+from eye_movement.utils import Gaze_Data, logger, can_use_3d_gaze_mapping, gaze_data_to_nslr_data
+from eye_movement.model.time_range import Time_Range
+from eye_movement.model.immutable_capture import Immutable_Capture
+from eye_movement.model.segment import Classified_Segment, Classified_Segment_Factory
+from stdlib_utils import sliceable_deque
+import numpy as np
+import nslr_hmm
+
+
+class Real_Time_Buffered_Detector:
+
+    def __init__(self, max_segment_count: int = 1, max_sample_count: int = 1000):
+        self._capture = None
+        self._gaze_data_buffer = sliceable_deque([], maxlen=max_sample_count)
+        self._segment_buffer = sliceable_deque([], maxlen=max_segment_count)
+        self._segment_factory = Classified_Segment_Factory()
+        self._is_gaze_buffer_classified: bool = True
+
+    def extend_gaze_data(self, gaze_data: Gaze_Data, capture: Immutable_Capture):
+        if not gaze_data:
+            return
+        self._capture = capture
+        self._gaze_data_buffer.extend(gaze_data)
+        self._is_gaze_buffer_classified = False
+
+    def segments_at_timestamp(self, target_timestamp: float) -> t.Iterable[Classified_Segment]:
+        self._update_classification()
+        return [segment for segment in self._segment_buffer if segment.time_range.contains(target_timestamp)]
+
+    def segments_in_time_range(self, target_range: Time_Range) -> t.Iterable[Classified_Segment]:
+        self._update_classification()
+        return [segment for segment in self._segment_buffer if segment.time_range.intersection(target_range)]
+
+    @staticmethod
+    def _segment_generator(capture: Immutable_Capture, gaze_data: Gaze_Data, factory_start_id: int = None):
+        # TODO: Merge this version with the one in offline_detection_task
+
+        if not gaze_data:
+            logger.warning("No data available to find fixations")
+            return
+
+        use_pupil = can_use_3d_gaze_mapping(gaze_data)
+
+        segment_factory = Classified_Segment_Factory(start_id=factory_start_id)
+
+        gaze_time = np.array([gp["timestamp"] for gp in gaze_data])
+
+        eye_positions = gaze_data_to_nslr_data(
+            capture, gaze_data, use_pupil=use_pupil
+        )
+
+        gaze_classification, segmentation, segment_classification = nslr_hmm.classify_gaze(
+            gaze_time, eye_positions
+        )
+
+        for i, nslr_segment in enumerate(segmentation.segments):
+
+            nslr_segment_class = segment_classification[i]
+
+            segment = segment_factory.create_segment(
+                gaze_data=gaze_data,
+                gaze_time=gaze_time,
+                use_pupil=use_pupil,
+                nslr_segment=nslr_segment,
+                nslr_segment_class=nslr_segment_class,
+            )
+
+            if not segment:
+                continue
+
+            yield segment
+
+    def _update_classification(self):
+        if self._is_gaze_buffer_classified:
+            return
+
+        factory_start_id = self._segment_buffer[0].id if len(self._segment_buffer) > 0 else None
+
+        segment_generator = type(self)._segment_generator(
+            capture=self._capture,
+            gaze_data=self._gaze_data_buffer,
+            factory_start_id=factory_start_id
+        )
+        new_segments = list(segment_generator)
+
+        # Update segment buffer by removing old segments and pushing new ones
+        self._segment_buffer.clear()
+        self._segment_buffer.extend(new_segments)
+
+        # Update gaze data buffer by removing any datapoints that precede the first classified segment
+        gaze_time_buffer = [gp["timestamp"] for gp in self._gaze_data_buffer]
+        start_timestamp = self._segment_buffer[0].start_frame_timestamp
+        i = bisect.bisect_left(gaze_time_buffer, start_timestamp)
+        self._gaze_data_buffer = self._gaze_data_buffer[i:]
+
+        # Mark current gaze data buffer as classified
+        self._is_gaze_buffer_classified = True
