+++ pupil_src/shared_modules/eye_movement_detector.py
@@ -48,222 +48,6 @@ from pyglui.cygl.utils import RGBA, draw_circle
 from pyglui.pyfontstash import fontstash
 
 
-Eye_Movement_Generator_Yield = typing.Tuple[
-    str, typing.Optional[MsgPack_Serialized_Segment]
-]
-
-
-Eye_Movement_Generator = typing.Generator[Eye_Movement_Generator_Yield, None, None]
-
-
-@typing.no_type_check
-def eye_movement_detection_generator(
-    capture: Immutable_Capture, gaze_data: Gaze_Data, factory_start_id: int = None
-) -> Eye_Movement_Generator:
-
-    def serialized_dict(datum):
-        if type(datum) is dict:
-            return fm.Serialized_Dict(python_dict=datum)
-        elif type(datum) is bytes:
-            return fm.Serialized_Dict(msgpack_bytes=datum)
-        else:
-            raise ValueError("Unsupported gaze datum type: {}.".format(type(datum)))
-
-    yield "Preparing gaze data...", ()
-    gaze_data = [ serialized_dict(datum) for datum in gaze_data ]
-
-    if not gaze_data:
-        logger.warning("No data available to find fixations")
-        yield "Segmentation complete", ()
-        return
-
-    use_pupil = can_use_3d_gaze_mapping(gaze_data)
-
-    segment_factory = Classified_Segment_Factory(start_id=factory_start_id)
-
-    gaze_time = np.array([gp["timestamp"] for gp in gaze_data])
-
-    yield "Processing {} gaze data...".format("3d" if use_pupil else "2d"), ()
-    eye_positions = preprocess_eye_movement_data(
-        capture, gaze_data, use_pupil=use_pupil
-    )
-
-    yield "Classifying {} gaze data...".format("3d" if use_pupil else "2d"), ()
-    gaze_classification, segmentation, segment_classification = nslr_hmm.classify_gaze(
-        gaze_time, eye_positions
-    )
-
-    # `gaze_classification` holds the classification for each gaze datum.
-
-    yield "Detecting segmentation events...", ()
-    for i, nslr_segment in enumerate(segmentation.segments):
-
-        nslr_segment_class = segment_classification[i]
-
-        segment = segment_factory.create_segment(
-            gaze_data=gaze_data,
-            gaze_time=gaze_time,
-            use_pupil=use_pupil,
-            nslr_segment=nslr_segment,
-            nslr_segment_class=nslr_segment_class,
-        )
-
-        if not segment:
-            continue
-
-        serialized = segment.to_msgpack()
-
-        yield "Detecting segmentation events...", serialized
-
-    yield "Segmentation complete", ()
-
-
-GFY = typing.TypeVar("GFY")  # Generator function yield type
-GFS = typing.TypeVar("GFS")  # Generator function send type
-GFR = typing.TypeVar("GFR")  # Generator function return type
-
-On_Started_Observer = typing.Callable[[], None]
-On_Yield_Observer = typing.Callable[[GFY], None]
-On_Completed_Observer = typing.Callable[[GFR], None]
-On_Ended = typing.Callable[[], None]
-On_Exception = typing.Callable[[Exception], None]
-On_Canceled_Or_Killed = typing.Callable[[], None]
-
-Eye_Movement_Detection_Task_Generator = typing.Generator[GFY, GFS, GFR]
-
-
-class Eye_Movement_Detection_Task(
-    BackgroundGeneratorFunction, typing.Generic[GFY, GFS, GFR]
-):
-    def __init__(
-        self,
-        name: str = "Segmentation",
-        generator_function: Eye_Movement_Detection_Task_Generator = eye_movement_detection_generator,
-        pass_shared_memory: bool = False,
-        args: GFS = None,
-        patches: typing.Iterable[typing.Type[Patch]] = tuple(),
-    ):
-        # TODO: A typed generic subclass of `BackgroundGeneratorFunction` can be added to `tasklib.background.task.py`
-        super().__init__(
-            name=name,
-            generator_function=generator_function,
-            pass_shared_memory=pass_shared_memory,
-            args=args,
-            kwargs={},
-            patches=patches,
-        )
-
-    def add_observers(
-        self,
-        on_started: typing.Optional[On_Started_Observer] = None,
-        on_yield: typing.Optional[On_Yield_Observer] = None,
-        on_completed: typing.Optional[On_Completed_Observer] = None,
-        on_ended: typing.Optional[On_Ended] = None,
-        on_exception: typing.Optional[On_Exception] = None,
-        on_canceled_or_killed: typing.Optional[On_Canceled_Or_Killed] = None,
-    ):
-        # TODO: A type-erased version of this method can be aaded to `TaskInterface`
-        if on_started:
-            self.add_observer("on_started", on_started)
-        if on_yield:
-            self.add_observer("on_yield", on_yield)
-        if on_completed:
-            self.add_observer("on_completed", on_completed)
-        if on_ended:
-            self.add_observer("on_ended", on_ended)
-        if on_exception:
-            self.add_observer("on_exception", on_exception)
-        if on_canceled_or_killed:
-            self.add_observer("on_canceled_or_killed", on_canceled_or_killed)
-
-
-class Eye_Movement_Buffered_Detection():
-    def __init__(self, max_segment_count: int = 1, max_sample_count: int = 1000):
-        self._capture = None
-        self._gaze_data_buffer = sliceable_deque([], maxlen=max_sample_count)
-        self._segment_buffer = sliceable_deque([], maxlen=max_segment_count)
-        self._segment_factory = Classified_Segment_Factory()
-        self._is_gaze_buffer_classified: bool = True
-
-    def extend_gaze_data(self, gaze_data: Gaze_Data, capture: Immutable_Capture):
-        if not gaze_data:
-            return
-        self._capture = capture
-        self._gaze_data_buffer.extend(gaze_data)
-        self._is_gaze_buffer_classified = False
-
-    def segments_at_timestamp(self, target_timestamp: float) -> typing.Iterable[Classified_Segment]:
-        self._update_classification()
-        return [ segment for segment in self._segment_buffer if segment.time_range.contains(target_timestamp) ]
-
-    def segments_in_time_range(self, target_range: Time_Range) -> typing.Iterable[Classified_Segment]:
-        self._update_classification()
-        return [ segment for segment in self._segment_buffer if segment.time_range.intersection(target_range) ]
-
-    def _segment_generator(self, capture: Immutable_Capture, gaze_data: Gaze_Data, factory_start_id: int = None):
-
-        if not gaze_data:
-            logger.warning("No data available to find fixations")
-            return
-
-        use_pupil = can_use_3d_gaze_mapping(gaze_data)
-
-        segment_factory = Classified_Segment_Factory(start_id=factory_start_id)
-
-        gaze_time = np.array([gp["timestamp"] for gp in gaze_data])
-
-        eye_positions = preprocess_eye_movement_data(
-            capture, gaze_data, use_pupil=use_pupil
-        )
-
-        gaze_classification, segmentation, segment_classification = nslr_hmm.classify_gaze(
-            gaze_time, eye_positions
-        )
-
-        for i, nslr_segment in enumerate(segmentation.segments):
-
-            nslr_segment_class = segment_classification[i]
-
-            segment = segment_factory.create_segment(
-                gaze_data=gaze_data,
-                gaze_time=gaze_time,
-                use_pupil=use_pupil,
-                nslr_segment=nslr_segment,
-                nslr_segment_class=nslr_segment_class,
-            )
-
-            if not segment:
-                continue
-
-            yield segment
-
-    def _update_classification(self):
-        if self._is_gaze_buffer_classified:
-            return
-
-        factory_start_id = self._segment_buffer[0].id if len(self._segment_buffer) > 0 else None
-
-        segment_generator = self._segment_generator(
-            capture=self._capture,
-            gaze_data=self._gaze_data_buffer,
-            factory_start_id=factory_start_id
-        )
-        new_segments = list(segment_generator)
-
-        # Update segment buffer by removing old segments and pushing new ones
-        self._segment_buffer.clear()
-        self._segment_buffer.extend(new_segments)
-
-        # Update gaze data buffer by removing any datapoints that precede the first classified segment
-        gaze_time_buffer = [ gp["timestamp"] for gp in self._gaze_data_buffer ]
-        start_timestamp = self._segment_buffer[0].start_frame_timestamp
-        i = bisect.bisect_left(gaze_time_buffer, start_timestamp)
-        self._gaze_data_buffer = self._gaze_data_buffer[i:]
-
-        # Mark current gaze data buffer as classified
-        self._is_gaze_buffer_classified = True
-
-
 
 class Notification_Subject:
     SHOULD_RECALCULATE = "segmentation_detector.should_recalculate"
